final1

+++ src/main/java/com/example/pi_ease/DAO/Entities/Credit.java
@@ -19,7 +19,7 @@ public class Credit  implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ID_C")
-    int idC;
+    long idC;
 
     @Temporal(TemporalType.DATE)
     @Column(name = "demand_date")
