+++ src/main/java/com/example/pi_ease/DAO/Repositories/CreditRepo.java
@@ -5,8 +5,8 @@ import com.example.pi_ease.DAO.Entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface CreditRepo extends JpaRepository<Credit,Integer> {
-    Credit findByIdCAndUserCredit(int idc, User user);
+public interface CreditRepo extends JpaRepository<Credit,Long> {
+    Credit findByIdCAndUserCredit(long idc, User user);
 
 
 }
